+++ runtests.py
@@ -1,4 +1,8 @@
+import sys
+
+#make test TAGS="list"
 
 if __name__ == '__main__':
+    tags = sys.argv[1:]
     import stdnet
-    stdnet.runtests()
+    stdnet.runtests(tags = tags)